Fixed issues with blob deletion
In the REST version the code never reached the server because the endpoint was missing in the RestExtendedBlobs interface.

In the gRPC version the code failed to delete the blob because the function is called with the blobURL instead of blobId. The gRPC client now extracts the blobId before calling the operation.